[feat] 권한 홈 뷰로 이전
