+++ app/src/main/java/com/dgu/smartcareapp/presentation/home/HomeScreen.kt
@@ -1,8 +1,10 @@
 package com.dgu.smartcareapp.presentation.home
 
+import android.Manifest
 import android.annotation.SuppressLint
-import android.graphics.drawable.Icon
 import android.util.Log
+import androidx.activity.compose.rememberLauncherForActivityResult
+import androidx.activity.result.contract.ActivityResultContracts
 import androidx.compose.foundation.BorderStroke
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
@@ -21,6 +23,8 @@ import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.Card
 import androidx.compose.material.Icon
 import androidx.compose.material.IconButton
+import androidx.compose.material3.AlertDialog
+import androidx.compose.material3.Button
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.ExtendedFloatingActionButton
 import androidx.compose.material3.Scaffold
@@ -28,10 +32,12 @@ import androidx.compose.material3.Text
 import androidx.compose.material3.TopAppBar
 import androidx.compose.material3.TopAppBarDefaults
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
@@ -61,7 +67,7 @@ fun Home(
     todoViewModel: TodoViewModel = hiltViewModel()
 ) {
 //    todoViewModel.delete()
-
+    Permission()
     val todoLists by todoViewModel.todoList.collectAsState()
 
     Scaffold(
@@ -229,4 +235,47 @@ fun FloatingActionBtn(onClickTodo: () -> Unit) {
         contentColor = Color.White,
         shape = RoundedCornerShape(corner = CornerSize(50.dp))
     )
+}
+
+
+@Composable
+fun Permission() {
+    var showPermissionDialog by remember { mutableStateOf(false) }
+    val permissionsLauncher = rememberLauncherForActivityResult(
+        contract = ActivityResultContracts.RequestMultiplePermissions()
+    ) { permissions ->
+        when {
+            permissions.values.all { it } -> {
+                // 권한이 모두 승인되었을 때의 로직
+            }
+
+            else -> {
+                // 권한 거부에 대한 처리
+                showPermissionDialog = true
+            }
+        }
+    }
+    LaunchedEffect(key1 = true) {
+        permissionsLauncher.launch(
+            arrayOf(
+                Manifest.permission.SEND_SMS,
+                Manifest.permission.RECEIVE_SMS,
+                Manifest.permission.READ_SMS,
+                Manifest.permission.RECORD_AUDIO
+            )
+        )
+    }
+
+    if (showPermissionDialog) {
+        AlertDialog(
+            onDismissRequest = { showPermissionDialog = false },
+            confirmButton = {
+                Button(onClick = { showPermissionDialog = false }) {
+                    Text("확인")
+                }
+            },
+            title = { Text("권한 요청 필요") },
+            text = { Text("보호자에게 경고 메시지를 보내기 위해 SMS 및 음성 녹음 권한이 필요합니다.") }
+        )
+    }
 }
