+++ app/src/main/java/com/dgu/smartcareapp/presentation/safeword/SafeWordViewModel.kt
@@ -36,5 +36,4 @@ class SafeWordViewModel @Inject constructor(
             getAllSafeWords()
         }
     }
-
 }
